Bit of cleanup

+++ servoy_ngclient/src/com/servoy/j2db/server/ngclient/property/ComponentTypeSabloValue.java
@@ -211,7 +211,7 @@ public class ComponentTypeSabloValue implements ISmartPropertyValue
 			Object value = childComponent.getFormElement().getPropertyValue(handler);
 			if (value != null)
 			{
-				childComponent.add(handler, (Integer)value);
+				childComponent.add(handler, ((Integer)value).intValue());
 			}
 		}
